refactor(content-mapper): share LSP test support
